Relocate select and slice benchmarks to test_select_and_slice_perf.py
